requirements.txt
deepgram-sdk==3.7.7
mutagen
openai
pydantic
pydub
python-dotenv
requests
spacy
torch
transformers